move transaction_in_progress to execution
