+++ quadratic-core/src/controller/user_actions/cells.rs
@@ -1,7 +1,6 @@
-use super::super::{
-    transaction_in_progress::TransactionType, transaction_summary::TransactionSummary,
-    GridController,
-};
+use super::super::{transaction_summary::TransactionSummary, GridController};
+use crate::controller::execution::TransactionType;
+
 use crate::{
     grid::{CodeCellLanguage, SheetId},
     Rect, SheetPos, SheetRect,
