+++ quadratic-core/src/controller/user_actions/undo.rs
@@ -1,7 +1,5 @@
-use super::super::{
-    transaction_in_progress::TransactionType, transaction_summary::TransactionSummary,
-    GridController,
-};
+use super::super::{transaction_summary::TransactionSummary, GridController};
+use crate::controller::execution::TransactionType;
 
 impl GridController {
     pub fn has_undo(&self) -> bool {
